requirements.txt
edge-tts
moviepy
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
groq
python-dotenv
streamlit
pydub
Pillow
